+++ volmdlr/edges.py
@@ -14,10 +14,11 @@ import dessia_common.core as dc
 import matplotlib.patches
 import matplotlib.pyplot as plt
 import numpy as npy
+from numpy.typing import NDArray
 import plot_data.core as plot_data
 import plot_data.colors
 import scipy.integrate as scipy_integrate
-from scipy.optimize import least_squares
+from scipy.optimize import least_squares, minimize
 from geomdl import NURBS, BSpline
 
 from volmdlr.nurbs.operations import split_curve, decompose_curve
@@ -889,20 +890,20 @@ class BSplineCurve(Edge):
 
     def __init__(self,
                  degree: int,
-                 control_points: Union[List[volmdlr.Point2D], List[volmdlr.Point3D]],
-                 knot_multiplicities: List[int],
-                 knots: List[float],
-                 weights: List[float] = None,
+                 control_points: Union[List[volmdlr.Point2D], List[volmdlr.Point3D], NDArray],
+                 knot_multiplicities: Union[List[int], NDArray],
+                 knots: Union[List[float], NDArray],
+                 weights: Union[List[float], NDArray] = None,
                  name: str = ''):
-        self.ctrlpts = npy.array(control_points)
+        self.ctrlpts = npy.asarray(control_points)
         self.degree = degree
-        self.knots = nurbs_helpers.standardize_knot_vector(knots)
-        self.knot_multiplicities = knot_multiplicities
+        self.knots = npy.asarray(nurbs_helpers.standardize_knot_vector(knots))
+        self.knot_multiplicities = npy.asarray(knot_multiplicities, dtype=npy.int16)
         self.weights = weights
         self.ctrlptsw = None
         self.rational = False
         if self.weights is not None:
-            self.weights = npy.array(weights, dtype=npy.float64)
+            self.weights = npy.asarray(weights, dtype=npy.float64)
             self.rational = self.weights.any()
             if self.rational:
                 self.ctrlptsw = npy.hstack((self.ctrlpts * self.weights[:, npy.newaxis], self.weights[:, npy.newaxis]))
@@ -958,10 +959,7 @@ class BSplineCurve(Edge):
     def knotvector(self):
         """Return the knot vector."""
         if self._knotvector is None:
-            knot_vector = []
-            for knot, knot_mut in zip(self.knots, self.knot_multiplicities):
-                knot_vector.extend([knot] * knot_mut)
-            self._knotvector = knot_vector
+            self._knotvector = npy.repeat(self.knots, self.knot_multiplicities)
         return self._knotvector
 
     @property
@@ -993,11 +991,11 @@ class BSplineCurve(Edge):
             "knotvector": self.knotvector,
             "size": self.ctrlpts.shape[0],
             "sample_size": self.sample_size,
-            "rational": not (self.weights is None),
+            "rational": self.rational,
             "dimension": 3 if self.__class__.__name__[-2:] == "3D" else 2,
             "precision": 18
         }
-        if self.weights is not None:
+        if self.rational:
             datadict["control_points"] = self.ctrlptsw
         else:
             datadict["control_points"] = self.ctrlpts
@@ -1068,10 +1066,20 @@ class BSplineCurve(Edge):
         """
         return self.knotvector[self.degree], self.knotvector[-(self.degree + 1)]
 
-    def copy(self, *args, **kwargs):
-        """Returns a copy of the instance."""
-        return self.__class__(self.degree, self.control_points, self.knot_multiplicities,
-                              self.knots, name=self.name + "_copy")
+    def copy(self, deep: bool = True, **kwargs):
+        """
+        Returns a copy of the instance.
+
+        :param deep: If False, perform a shallow copy. If True, perform a deep copy.
+        """
+        if deep:
+            weights = None
+            if self.rational:
+                weights = self.weights.copy()
+            return self.__class__(self.degree, self.control_points, self.knot_multiplicities.copy(),
+                                  self.knots.copy(), weights, name=self.name + "_copy")
+        return self.__class__(self.degree, self.ctrlpts, self.knot_multiplicities,
+                              self.knots, self.weights, name=self.name + "_copy")
 
     def to_geomdl(self):
         """Converts the BSpline curve into a geomdl curve."""
@@ -1093,9 +1101,11 @@ class BSplineCurve(Edge):
         dict_ = self.base_dict()
         dict_['degree'] = self.degree
         dict_['control_points'] = [point.to_dict() for point in self.control_points]
-        dict_['knot_multiplicities'] = self.knot_multiplicities
-        dict_['knots'] = self.knots
-        dict_['weights'] = self.weights
+        dict_['knot_multiplicities'] = self.knot_multiplicities.tolist()
+        dict_['knots'] = self.knots.tolist()
+        dict_['weights'] = None
+        if self.rational:
+            dict_['weights'] = self.weights.tolist()
         return dict_
 
     def decompose(self, return_params: bool = False):
@@ -1183,10 +1193,10 @@ class BSplineCurve(Edge):
             "knotvector": self.knotvector,
             "size": self.ctrlpts.shape[0],
             "sample_size": self.sample_size,
-            "rational": True if self.weights is not None else False,
+            "rational": self.rational,
             "dimension": 3 if vector_name == "Vector3D" else 2,
         }
-        if self.weights is not None:
+        if self.rational:
             datadict["control_points"] = self.ctrlptsw
         else:
             datadict["control_points"] = self.ctrlpts
@@ -1373,9 +1383,10 @@ class BSplineCurve(Edge):
         if point.is_close(self.end):
             return self.length()
         length = self.length()
-        point_array = npy.array(list(point), dtype=npy.float64)
+        point_array = npy.asarray(point)
         distances = npy.linalg.norm(self._eval_points - point_array, axis=1)
-        index = npy.argmin(distances)
+        indexes = npy.argsort(distances)
+        index = indexes[0]
         u_min, u_max = self.domain
         u0 = u_min + index * (u_max - u_min) / (self.sample_size - 1)
         u, convergence_sucess = self.point_invertion(u0, point)
@@ -1388,19 +1399,13 @@ class BSplineCurve(Edge):
         def evaluate_point_distance(u_param):
             return (point - self.evaluate_single(u_param)).norm()
         results = [(abscissa, evaluate_point_distance(u))]
-        initial_condition_list = npy.linspace(u_min, u_max, 10).tolist()
-        initial_condition_list.sort(key=evaluate_point_distance)
+        initial_condition_list = [u_min + index * (u_max - u_min) / (self.sample_size - 1) for index in indexes[:3]]
         for u0 in initial_condition_list:
-            u, convergence_sucess = self.point_invertion(u0, point)
-            if u_min != 0 or u_max != 1.0:
-                u = (u - u_min) / (u_max - u_min)
-            abscissa = u * length
-            if convergence_sucess:  # sometimes we don't achieve convergence with a given initial guess
-                return float(abscissa)
-            dist = evaluate_point_distance(u)
-            if dist < tol:
-                return float(abscissa)
-            results.append((abscissa, dist))
+            res = minimize(evaluate_point_distance, npy.array(u0), bounds=[(u_min, u_max)])
+            if res.fun < 1e-6:
+                return float(res.x[0] * length)
+            abscissa = res.x[0] * length
+            results.append((abscissa, res.fun))
         result = min(results, key=lambda r: r[1])[0]
         return float(result)
 
@@ -4803,7 +4808,7 @@ class LineSegment3D(LineSegment):
         if self.point_belongs(line_intersection):
             return self._helper_intersecting_axis_plane_revolution(surface, distance_1, distance_2, angle)
         smaller_r, bigger_r = sorted([distance_1, distance_2])
-        if angle == volmdlr.TWO_PI:
+        if math.isclose(angle, volmdlr.TWO_PI, abs_tol=1e-6):
             return self._helper_plane_revolution_two_circles(surface, bigger_r, smaller_r)
         return self._helper_plane_revolution_arcs_and_lines(surface, bigger_r, smaller_r, angle)
 
@@ -5057,12 +5062,8 @@ class BSplineCurve3D(BSplineCurve):
         name = arguments[0][1:-1]
         degree = int(arguments[1])
         points = [object_dict[int(i[1:])] for i in arguments[2]]
-        lines = [LineSegment3D(pt1, pt2) for pt1, pt2 in zip(points[:-1], points[1:]) if not pt1.is_close(pt2)]
-        if lines and not points[0].is_close(points[-1]):
-            # quick fix. Real problem: Tolerance too low (1e-6 m = 0.001mm)
-            dir_vector = lines[0].unit_direction_vector()
-            if all(line.unit_direction_vector() == dir_vector for line in lines):
-                return LineSegment3D(points[0], points[-1])
+        if len(points) == 2 and degree == 1:
+            return LineSegment3D(points[0], points[-1])
 
         knot_multiplicities = [int(i) for i in arguments[6][1:-1].split(",")]
         knots = [float(i) for i in arguments[7][1:-1].split(",")]