+++ volmdlr/utils/step_reader.py
@@ -3,11 +3,10 @@ volmdlr utils for importing step files.
 """
 import re
 
-import numpy as npy
-
 import volmdlr
 import volmdlr.shells as vmshells
 from volmdlr import surfaces
+from volmdlr.geometry import get_transfer_matrix_from_basis
 
 
 def set_to_list(step_set):
@@ -728,13 +727,7 @@ def frame_map_closed_shell(closed_shells, item_defined_transformation_frames, sh
     for shell3d in closed_shells:
         basis_a = global_frame.basis()
         basis_b = transformed_frame.basis()
-        matrix_a = npy.array([[basis_a.vectors[0].x, basis_a.vectors[0].y, basis_a.vectors[0].z],
-                              [basis_a.vectors[1].x, basis_a.vectors[1].y, basis_a.vectors[1].z],
-                              [basis_a.vectors[2].x, basis_a.vectors[2].y, basis_a.vectors[2].z]])
-        matrix_b = npy.array([[basis_b.vectors[0].x, basis_b.vectors[0].y, basis_b.vectors[0].z],
-                              [basis_b.vectors[1].x, basis_b.vectors[1].y, basis_b.vectors[1].z],
-                              [basis_b.vectors[2].x, basis_b.vectors[2].y, basis_b.vectors[2].z]])
-        transfer_matrix = npy.linalg.solve(matrix_a, matrix_b)
+        transfer_matrix = get_transfer_matrix_from_basis(basis_a, basis_b)
         new_frame = volmdlr.Frame3D(transformed_frame.origin, volmdlr.Vector3D(*transfer_matrix[0]),
                                     volmdlr.Vector3D(*transfer_matrix[1]), volmdlr.Vector3D(*transfer_matrix[2]))
         new_closedshells.append(shell3d.frame_mapping(new_frame, 'old'))